Properly wrap LV2 searchpath on nix builds
For LV2 plugins to be visible, they need to be in the LV2 search path.
This patch allows a list of lv2 plugins to be passed as well as a flag
to enable a default list of plugins. To enable the default list, you
can build with:

nix-build --arg  defaultLv2Plugins true