CMake: Get rid of duplicate target_include_directories for Katai
When an include directory for a target is marked as PUBLIC, it will also
be added as to all targets that depend on it (i.e. if the target is a
library, all targets that this library is linked to via target_link_libraries
will also have this include directory). Hence, there's no need to set
this include directory twice.