Master Sync: preallocate buffers in engine thread func.
pickMaster gets called from the engine thread so we have to be careful to
avoid memory allocation.  Preallocating a constant number fixes this.

Also make kMaxNumberOfDecks a proper mixxx-global value, and check against
it when we add decks.  If someone one day invents a skin that uses 6 or 8
decks, this warning will get tripped (but shouldn't break anything
fundamental in Mixxx).