Adding Training mode
This adds an training mode. This should enable you to train with your
database and improve, but do not add your training to the play count. As
the play count should be related to the count that I have played this
song on events/broadcasts.

The design is done in a way that uses the Mixxx's Control System
functionality to comunicate between the gui and the actual counter
implementation.