scripts: Add serato heartbeat script
This script features an interactive prompt to simulate Serato mode on
controllers. It first sends the Serato SysEx messages to a Serato MIDI
controller, then continuously sends a special MIDI heartbeat message.
This is for controllers like the Roland DJ-505, that will fall back into
a generic mode with a reduced feature set if that's not the case.