Fix db inconsistencies of a fresh install
On a fresh install of Mixxx. I imported and analyzed some tracks into
the library. Thereafter, the sqlite database was stating it was
corrupted. Discovered while running an "PRAGMA integrity_check;" or
operations in foreign-key check enabled mode on the sqlite db.

Mainly the data was not corrupted, but the scheme was referencing to non
existing columns, tables or wrong tables. This is going to fix this
false positive corruption and might also be able to run mixxx in a mode
where it makes it requests to the db with foreign key checks enabled.