Roland DJ-505: Fix SYNC button behaviour
This fixes a nasty little bug that was discovered while testing #2376.
Before, pressing the SYNC button always caused beatsync, even on long
press. This caused sudden jumps/rate changes on the playing deck even
when the other deck did not have sync enabled and was not playing.

This also adds support for setting sync_master on a deck:
When sync is already enabled and the button is long pressed, sync_master
is enabled. The SYNC button indicates the master state with a blinking
LED.