master sync: differentiate soft from explicit master
SYNC_MASTER really contained two ideas: a master that the user had
explicitly chosen, or a master that Mixxx had chosen automatically
in order to make sync work.  It became clear that we needed
to differentiate the two, because there are cases where they need
to be treated differently.

So: If a user selects a master explicitly, that deck will remain
master until it is stopped. (If the deck stops and another sync
deck is playing, that deck would also just stop dead, which is
not a great behavior!).

Explicit master could be useful in the instance of, say, having
a beatmap track playing that you want to drive the bpm of the
other track no matter what.