EffectManifestParameter: track index in the EffectManifest
This ensures that EffectParameter always controls the correct
EngineEffectParameter. Before, EffectParameter took the manifest
parameter index as an argument to its constructor, which left room
for error when constructing EffectParameters in
EffectSlot::loadEffect. When I was working on parameter hiding
and rearrangement from EffectPresets (c8367b1), I got quite
confused for a while when I accidentally messed up the index I was
passing to the EffectParameter constructor.