remove unused features from EngineEffectParameter
EngineEffectParameters have absolute minimum and maximum limits
specified by the EffectParameterManifest. They cannot be changed
after construction. If setting custom minimum and maximum values
for parameters is implemented, that will be handled by
EffectParameter in the main thread and those must be narrower than
the absolute limits specified in the manifest, or the DSP logic
of effects may not work.