controllers/controllerengine: Always show the backtrace on script errors
Previously the backtrace was only added to the error details if
the `--controllerDebug` flag was passed to Mixxx. Since the error
details are only displayed if the user explicitly uncollapsed the
"Details" expander and non-technical users won't be able to make use of
the engine's error message and line number anyway, we can just show it
unconditionally. This might even result in more helpful bug reports.