removed the use of file_bpm form the engine, streamline update order.
Depending on the event different slots of bpmconro and synccontrol where
called in different order. This leads to various hard to tack issues
because the cal heriarchy is unclear when reading signal slot connections.

Now bpmconrol is always updated before synccontrol and we onbly look to Beats.
Loading a track and changing beats are treated the same.