preferences/dialog/dlgprefdeck: Tick "Down increases Speed" by default
With this change, Mixxx is now consistent with almost every other DJ
application or hardware out there.

This also renames the internal variable, be "inverted" depends on your
perspective - is "Down increases speed" inverted because "low fader" =
"low speed"? Or is "Up increases speed" inverted because it's the
opposite of what the rest of the DJ world does?
The new name should be less ambiguous.

Zulip Discussion:
https://mixxx.zulipchat.com/#narrow/stream/109122-general/topic/Tempo.20Slider.20default.20behavior