fix woverview playpos not being updated when clicking marker
When a marker was clicked previously when the track was paused,
the ruler jumped to the marker but the playposition of the
overview waveform (which influences the gray area) was not
getting updated. The actual playposition was updated however.
This left the playposition of the overview and the real one
out-of-sync.