waveform/renderers/waveformmarkset: Draw hotcue mark with lowest index on top
Currently, hotcue marks on the same position are drawn over each other
and only the hotcue mark with the highest index is shown. Since not all
hotcue indices might be mapped on the controller (or accessible via the
skin) it makes sense to draw mark with the lowest index on top to
increase the chance that a user can jump to it via GUI or performance
pads.

Zulip Discussion:
https://mixxx.zulipchat.com/#narrow/stream/109171-development/topic/hotcue.20markers.20drawn.20over.20each.20other