Rename scripts/ directory to tools/
We also have a "script" directory (without the "s"), which makes it confusing
and harder to select the correct directory with shell completion.

The "tools" name fits better, since the directory contains tools that are/were
helpful during mixxx development (these tools are all scripts - but that
is of secondary importance and also coincidence).