Fix Traktor F1 output packet mapping
HIDPacket.addControl() expects its offset to start at 0 and applies to
the packet payload (which does not take the header into account).

The Traktor F1 expects the output report to be 81 bytes long with its
first byte (the header) set to 0x80. Moreover, sending a packet larger
than 81 bytes seems to stall the device on Linux. At this point the
device needs to be unplugged/plugged in order to accept new packets.

Offsetting the current values by one fixes the output LEDs and
avoids stalling the device on Linux.