remove "controller engine version" concept
This was introduced in PR #1795 when the migration to QJSEngine was
planned for a release when the latest Ubuntu version shipped Qt 5.9,
which did not support ES7. The purpose was to show users an error when
they loaded a script using ES7 if they were using a Qt version before
5.12. Since Mixxx 2.3 took so long, the QJSEngine migration will be
released when the latest Ubuntu LTS (20.04) ships Qt 5.12, so we do not
need to introduce this "controller engine version" concept anymore.

Mixxx will still build and run with Qt < 5.12 at the small cost of
keeping only one #if in ControllerEngine::throwJSError, but mappings
using ES7, which now includes any mapping using the Components library,
will not work.