downbeats: add persistence
Persistence works fine for both beatgrid and beatmap.
But the deserialization result for beatmap's downbeat index is
overriden by the analyzer thread which is currently incapable
of generating downbeat information.

The downbeat restore on mixxx reopen works fine for beatgrid though.

This behavior can be fixed with the #2512 merge or with downbeat
detection.