rename Master output to Main output in user visible strings
There is still more work to remove this terminology from the code.
"Master" is still used in Shade but I do not know how to change
that because it is in a raster graphic in
res/skins/Shade/style/style_bg_mixer_panel_main.png and I do not
know what font to use to replace it.