colorpaletteeditor: remove last column
this commit removes the last column of the colorpaletteeditor
which was solely used for cosmetic purposes. The cells of that
column were editable even though they were meaningless and thus
could cause confusion for the user. This commit removes the
cosmetic column and tells the last header section to fill
the space automatically.