LegacySkinParser: Persist skin-created COs even on skin changes
Up until now, Skin-created COs were parented to a skin widget, which
means that they will be deleted when the skin is unloaded and the skin
widget is destroyed. This sounds like a good idea, but it's actually
quite problematic.

Controller mappings may connect to these COs, and as they are not
reloaded when the skin is changed these ControlProxy objects will still
hold a reference to the internal ControlDoublePrivate, even though the
actual ControlObject has been deleted.
When switching to a skin that attempts to create a CO that has been
destroyed earlier, this causes a DEBUG_ASSERT.

This commit removes the parenting of skin-created control objects.
Instead, all skin-created COs are put into a global set of skin COs, and
its contents will only be destroyed in MixxxMainWindow::finalize(), i.e.
when Mixxx is exited. That way, all skin-created COs persist accross
skin changes and skins can reuse COs created by other skins.

This is definitely not a pretty "solution", but it does the job:
It fixes the DEBUG_ASSERT that was thrown on skin change when a
controller mapping was connected a CO that was destroyed and the new
skin tries to create a new CO under the same key. The controller mapping
keeps working even after the skin change.

Related Zulip discussion: https://mixxx.zulipchat.com/#narrow/stream/109171-development/topic/DEBUG_ASSERT.20on.20skin.20change