engine/enginebuffer: Move CO setting into locking scope
I sometimes see random segfauls in the enginebuffer tests. According to
the backtrace it happens when calling m_pPlayButton->set(v) in the
EngineBuffer::slotTrackLoading() method.

I'm not really familiar with the code but I think it's possible that
this happens because the settings of `m_pTrackLoaded` and `m_pTrackSamples`
are not inside the lock. If the track loads very fast, then
`slotTrackLoaded`/`slotLoadTrackFailed` could be called before
`slotTrackLoading` is finished, which could lead to inconstencies of
the `m_pTrackSamples` value.