ControllerManager: poll every 1ms on Linux
Polling at 5ms causes a noticable lag with the GUI. If the audio
buffer is set low, it also produces a lag with the audio. The
bug report ( https://bugs.launchpad.net/mixxx/+bug/990992 )
referenced for why this was increased to 5 ms on Linux is 8 years
old. The root cause of that bug was not identified, but increasing
the controller polling latency to 5ms worked around it. After
setting the polling latency back to 1ms, I can not reproduce the
bug with high CPU usage on modern Linux (Fedora 32 running kernel
5.4.39-rt23.fc32.x86_64 with realtime patch set).