Fix save order, add round button
Detect when the BPM value was changed by the user and save
the BPM value in such cases on apply.
This allows to unlock, change the BPM value, lock again and then press
apply and the value is saved.
Previously the user had to unlock, change the BPM, press apply, then
lock the track again and then apply, which is a not very intuitive workflow.

Add round button next to BPM spin box to round the BPM number to
the next integer value. Very handy on 2/3 multipliers.