Use signal to initiate state save
Use a signal in Library or LibraryFeature objects to
save the scroll position of the trackmodel currently shown.
The key to store the scroll position is no longer defined by the
model pointer but a value provided by the model itself.
This way, the same model can be reused for different views like
the crates feature does.