Fix resize problem, update history behaviour
Update internal metrics to calculate button positions on font changes.

Move the new entry to the top of the list when an old entry was found.
This way the search behaves more like expected.
Avoid duplicates when Enter is pressed.
Save new query term if keydown is pressed or popup opened.