EffectChainPresetManager: refactor loading of preset lists
This works around an annoying situation when switching between
Mixxx 2.4 and Mixxx 2.1 - 2.3 where the older version would
save an effects.xml file without the ordered lists of chain
presets, so all the preset files from the user settings folder
would be added to both the regular and QuickEffect preset lists.

Now, if effects.xml does not provide those ordered lists, they
are generated from defaults. Then, if there are any presets
in the user settings folder that are not in either list, they
are prepended to both lists.