Fix broken reload screen logic
The original code fails on multi-screen setups and places the window
in between the displays. Unfortunately, there seems to be a bug in QT
which causes screen() to return stale data if the window is not moved
or clicked on after the reload. If you use shortcuts to reload
the skin, saving the screen does not work either.

Instead of fancy logic, simply safe the geometry and restore it
after reload, so the window will not move around when testing.