Slimdown moc_compilation.cpp
The moc_compilation file became unbearable large, easily taking 10 minutes
to compile whenever a Q_OBJECT changes.
AUTOMOC scans for <basename>_moc.cpp includes and does not add the file
to the moc_compilation whenever it is included in the propper cpp file.

After this patch, moc_compilation.cpp is empty and the mocs are compiled
with the approptial cpp, increasing recompile speeds by a factor of 10 when
just a header file changed.