Library: overhaul setting font and row height
Previously, there was an artificial lower limit for the row height (31px
for me) that would ignore lower values of the row height spinbox.

Now, the new font height is used as lower limit for the row height.
When changing the font (size) the row height is automatically adjusted
by using the previous font height/row height ratio.