CMake: do not use /OPT:ICF link option on Windows
This caused a linker warning:
LINK : warning LNK4075: ignoring '/INCREMENTAL' due to '/OPT:ICF' specification

Also, it makes debugging confusing. From
https://docs.microsoft.com/en-us/cpp/build/reference/opt-optimizations?view=msvc-160

because /OPT:ICF can merge identical data or functions, it can
change the function names that appear in stack traces. It can also
make it impossible to set breakpoints in certain functions or to
examine some data in the debugger, and can take you into
unexpected functions when you single-step through your code. The
behavior of the code is identical, but the debugger presentation
can be very confusing. Therefore, we do not recommend that you use
/OPT:ICF in debug builds unless the advantages of smaller code
outweigh these disadvantages.