CMake: Use default_option instead of cmake_dependent_option for defaults
Before, we often used `cmake_dependent_option` for options depending on
the availability of dependencies. However, this is not the proper way to
use that macro, because if the dependencies is missing, it will forcibly
set the option to off and not cause a warning if the user tries to set
it to on.

`cmake_dependent_option` should only be used for disabling options based
on the platform, where that option will never be available, not based on
dependencies that may be present or not.

Instead, this adds a new macro `default_option` that allows to provide a
default depending on the value of other variables/options. The user may
still override that default and will run into an error message in that
case. From the user's perspective this is preferable instead of silently
dropping the option, because then the user will only notice afterwards
that some functionality is missing (despite setting the corresponding
feature flag `-DUSE_FOO=ON` on the command line).