SeratoTags: Use SignalInfo/Duration instead of StreamInfo
For FLAC files, the StreamInfo might be invalid because the bitrate is
invalid, because it is read by TagLib, not the sound source.  We don't
actually need the Bitrate for exporting Serato Tags, but the signal info
alone does not suffice either, because we also need the duration to
serialize the beatgrid. Hence, instead of using the StreamInfo wrapper
(which also has the irrelevant Bitrate member) we use SignalInfo and
Duration directly.