Track: Do not discard all cue types on import unnecessarily (lp1920087)
When importing Serato Metadata from file tags, the current hotcues are
overwritten with the hotcues from Serato. This is correct behavior. But
in addition to that, the main cue and the intro/outro cues are reset as
well. This is unnecessary because Serato metadata doesn't store this
information anyway.

Launchpad issue: https://bugs.launchpad.net/mixxx/+bug/1920087