Timer: Added the initial requirements for the timer.
This commit intrdouces an unactionable button in the toolbar
labelled timer. It is expected to support a right click
which will trigger a menu and host a stopwatch timer.

Note:- This commit is a WIP and doesn't function properly.
This commit will be referenced in future commits to mark
the completion of functionalities.