Translations: Use POSIX-style locale identifiers
Switch to using an underscore to separate ISO-639-1 language and
ISO-3166-1 country codes (instead of a dash).

Also removes the `lang_map` setting from the Transifex client
configuration that replaces the underscore with a dash:
https://docs.transifex.com/client/client-configuration#language-mappings

This fixes the broken loading of translations that used a dash in the
filename, because this is not supported by `QTranslator::load()`.
See https://doc.qt.io/qt-5/qtranslator.html#load-1 for details.