DlgTrackInfo: Use TrackRecord as model
Save all edits temporarily in a local TrackRecord and replace them in the
Track object all at once on apply. This is required to keep the properties
in sync and valid while editing. Validation should not be done by the
dialog itself but by TrackRecord which serves as the model that contains
the domain logic.

This change becomes crucial when properties dependent on each other,
e.g. when storing genres both as multiple, custom tags and redundantly
formatted in a text property. Those properties need to be synchronized
mutually with each other.