Install app icon in XDG hicolor icon theme
The application icon is a scalable icon (SVG) rather than a pixmap, so
it does not fit well in that directory. Instead, install it in the
hicolor XDG icon theme: this way it can be properly loaded by XDG menus
in the currently set XDG icon theme, without looking in the legacy
pixmaps location.