Rename app SVG icon to 'mixxx'
As suggested by Daniel Schürmann, rename the main application icon to
'mixxx', updating all the reference to it in both sources and additional
metadata.

There are few XDG icon themes that provide an icon for mixxx, and they
all seem to provide already a 'mixxx' icon, so they will not see any
missing icon.