raspi: Fix QT onscreen keyboard masking.
Raspberry pi users (and others) may want to use the QT virtual keyboard
to do text input (like, in the Search bar).  This can be done by
setting the environment variable QT_IM_MODULE="qtvirtualkeyboard".

This patch fixes the opacity of the keyboard, which doesn't have its
masking set correctly.  This patch should be safe for non-raspi
situations -- when the input mode changes it specifically looks for
the virtual keyboard window and modifies it.