[DDJ 200] Get knobs/sliders position at init
This change add a line based on DDJ 400 mapping with a little of reverse engineering in my DDJ 200 controller to start mixxx with all knobs and sliders at same position as the controller.

I intercepted messages between rekordbox and my DJJ 200 with wireshark to find some command similar to the command I found at DDJ 400 mapping and found this. It's not documented in DDJ 200 midi manual but works like a charm.