Add support for editing ReplayGain values.
Sometimes the replaygain value is a bit too aggressive  -- maybe it's
technically correct, but the value isn't helpful for making mixes.
Add support for editing the value using the db adjustment string as
displayed in the track table.

Tested with several problem tracks, as well as tracks without any
replaygain value. If the user wants to restore the replaygain deteced
value, they can clear the replaygain value and reanalyze.