Fix wrong base tag used for deployment and displayed in About dialog
By default, `git describe` shows the number of commits from the latest
tag in the git history. In our case we don't want that, because this can
cause problems:

            2.4-alpha              Expected: 2.4-alpha-N-abcdef123
    main ---+X---------------+---> Actual: 2.3.0-M-abcdef456
             \              /
              X------------X--->
             2.3-beta     2.3.0

Hence, we need to add the `--first-parent` flag to git describe, so that
it only follows the first parent of a merge and does not pull in tags
from another branch.

Fixes mixxxdj/website#248.

See the issue for details: https://github.com/mixxxdj/website/issues/248