Beats: Remove unnecessary capabilities
We currently have two beats implementations (`BeatGrid` and `BeatMap`)
and both support scaling and translating. Hence, it's unnecessary to
maintain these flags. In fact, the `BEATSCAP_SCALE` capability wasn't
even checked in the calling code but nobody noticed because both support
it anyway.

The `BEATSCAP_ADDREMOVE` and `BEATSCAP_MOVEBEAT` capabilities are only
supported by the `BeatMap`, but we don't have any code that makes use of
them. Therefore, these can be removed as well.