ControlIndicatorTimer: Add support for legacy vsyncthread mode
Due to concerns that a regular QTimer that fires during vsync could
cause jerking legacy waveforms, this implements a way to use the
legacy vsync thread/GuiTick infrastructure instead of a `QTimer` in the
`ControlIndicatorTimer` (although I'm still not convinced that
this actually makes a difference).

This can be toggled explicitly using `setLegacyVsyncEnabled` when
switching between a legacy and a QML skin (the latter has no vsync
thread, so a regular timer needs to be used).

Relevant discussion can be found here:
https://github.com/mixxxdj/mixxx/pull/4157#issuecomment-888945348