ControllerScriptEngineBase: only shutdown controller engine if it exists
When loading a faulty controller script, the engine might gets destroyed
when no script is using it. So when reloading when the file changes,
trying to shutdown an engine that doesn't exist triggers a debug assert
or crashes mixxx. So in that case, skip the shutdown procedure and just
re-initialize it instead.