DB: Update track metadata from file tags if outdated
When loading a track from the database check if the underlying file
has been modified. If the time stamp is newer than the timestamp of
the last synchronization then update track metadata from file tags.

But only if export of track metadata into file tags has been enabled
by the user to reduce surprising side-effects to a minimum. Introducing
a new configuration option for this feature has deliberately been avoided
to reduce the number of possible configurations. Either you want to
keep your file tags synchronized or not, no exceptions!

# Conflicts:
#	src/sources/metadatasourcetaglib.cpp