BeatsImporter: Return Beats object instead of frame position vector
This makes the BeatsImporter return an actual Beats object instead of a
vector of frame positions. The main reason for this change is that it
will avoid the unnecessary intermediate step of converting to/from raw
frame positions when the new beats implementation is ready.

Additionally, this prevents wrong usage by making it impossible to
passing a `SignalInfo` to import method and then use a different sample
rate for creating the beats object (which would invalidate all
positions).

Theoretically, this also allows to return a `BeatGrid` instead of a
`BeatMap`, although we don't make use of this possibility and the
distinction hopefully becomes obsolete soon anyway.