Fix race condition in ReplayGain test
This test caused a race because it called slotLoadTrack instead of loadTrack, so it did not wait for the track to actually load. The test was quitting before the caching reader had a chance to finish setting up, so when the test was getting torn down there was a chance that the reader would try to access the objects while they were being destructed.

Because the test infrastructure has its own destruction code, it's not clear if this crash could occur in regular code.

Fixes https://bugs.launchpad.net/mixxx/+bug/1940589