EffectChainPresetListModel: fix validation of dropped text
Before, the dropped text was checked if it was in the user's set
of presets for the standard effect chains. That broke dropping
presets in the QuickEffect list that were not in the standard
effect chain preset list and also broke dragging and dropping
presets from the QuickEffect list to the standard chain preset
list.