remove library sidebar from QML
As discussed on Zulip, we will not be continuing to work on fitting
the legacy LibraryFeatures into QML. The simple ListView for the
Tracks feature remains as a crude hack useful for developing QML
until we replace it with a new QML GUI for Aoide. The sidebar
was not functional anyway; it only showed the data with no way to
make use of it.
https://mixxx.zulipchat.com/#narrow/stream/109171-development/topic/QML.20library.20GUI/near/257370390