Beats: Refactor getBpm() into getBpmInRange(startPos, endPos)
This lays some groundwork for a future, sparse representation of beats
in a track.

At the moment, we have two kinds of beats implementations:

1. The `BeatMap` representation is simply a list of every single beat
   position in a track. Hence, changing tempos are supported, but memory
   intensive to work with.
2. The `BeatGrid` representation is actually sparse (only stores the
   offset and bpm instead of actual beat positions), but only supports a
   single constant tempo.

Both are not sufficient. A future implementation should be both sparse
and support multiple tempos per track. To make this possible, we need to
know the track duration for calculating the average BPM of a track.

There are two ways to handle this: We could either store the track
duration in a beats object, or we change the `getBpm()` method signature
to require passing the position range of the track that we want to know
the BPM for.

This commit implements the latter solution, because it's fully
backwards-compatible and - in contrast to the former - can be
implemented *without* modifying the protobuf format of the `BeatGrid`
class (which currently doesn't store the track end position).

Related Zulip discussion:
https://mixxx.zulipchat.com/#narrow/stream/109171-development/topic/beat.20grid.20revamp.20re.3A.20pr.20.232961/near/251702242