Beats: Skip additional marker insertion if beats are beatgrid-compatible
Without this patch, every beats instance created from a beats vector
would be stored a `BeatMap` in the database, even if a `BeatGrid` would
suffice.

This also "fixes" the broken `EngineSyncTest.HalfDoubleConsistency` test
which suffered from a minor rounding issue in
`Beats::getBpmAroundPosition`, which is causes by the initial play
position of an `EngineBuffer` being set to
`std::numeric_limits<double>::lowest()`, which leads to slightly
inaccurate floating point results in `getBpmAroundPosition()`.

Because this commit makes the beats instance used in the test have a
constant tempo, this "fixes" the issue because the
`getBpmAroundPosition()` can exit early and just return the (accurate)
end marker BPM.