.pre-commit-config.yaml: disallow commiting to version branches
before this commit, it was possible to accidentally commit to
the stable version branches (such as 2.3). If this was done
by a member with push access, they could push untested commits
into production on accident. If this was done by a git newbie,
they would mess up their local branch and deal with conflicts
when pulling their branch later. If commiting to the branch
is justified, the check can be bypassed by
`SKIP=no-commit-to-branch git commit`