test: Fix tempfile handling and remove unnecessary utility functions
The QTemporaryFile/QTemporaryDir classes are very convenient and fit our
use cases well, because they can create a temporary file/dir that
automatically gets deleted when the object goes out of scope.

Instead of using these classes directly, we were using wrappers (namely
`generateTemporaryFileName` and `createEmptyTemporaryFile`) that created
a temp file and then just returned its name. This renders the
QTemporaryFile auto-removal mechanism useless because the object goes
out of scope before writing it.

To mitigate this, we used a custom `FileRemover` class to work around the
broken tempfile auto-removal. This is error-prone and we actually forgot
to add it in some tests:
https://mixxx.zulipchat.com/#narrow/stream/247620-development-help/topic/git.20ignore.20generated.20test.20files.3F

Another major issue with the old API was that we often passed a
*filename* to `generateTemporaryFileName` instead of a filename
*template*. The distinction is important, because the latter may append
random characters to the passed string:

> If the file template contains XXXXXX that will automatically be
> replaced with the unique part of the filename, otherwise a filename
> will be determined automatically based on the static portion specified.
>
> Source: https://doc.qt.io/qt-5/qtemporaryfile.html#setFileTemplate

Hence, the generated temporary file names weren't suitable for tests
where the filename was important, such as
SoundSourceProxyTest.getTypeFromFile where were didn't actually check
wappens in cases where the file suffix is missing, misleading or empty,
but unknowingly always checked the case that the suffix is unknown (i.e.
instead of testing `file_with_no_file_suffix` we checked
`file_with_no_file_suffix.HPqSHg`).

This commit removes the APIs that break our tests and replaces them with
QTemporaryDir/QTemporaryFile instead, so that the tests work as expected
and don't leave temporary files behind.